Add Redis Broker
This adds broker implementation that uses redis to send data between
publishers and consumers. It uses the reliable message pattern for redis
using BRPOPLPUSH to be careful to not orphan data within a worker before
it is acknowledged.

All brokers checker on all other brokers for ones that have become dead
for any reason. If a dead broken is detected, it's unacked messages are
reclaimed back into the original queue to be handled. This behavior is
automatic and does not require an additional cleanup broker.